database dump file en images gedownload en troep weggegooid en de scraper aangepast

+++ offline/Rijksmonumenten Scraper/src/org/Monumentzo/RijksmonumtenScraper/ImageScraper.java
@@ -13,6 +13,15 @@ public class ImageScraper {
 	public static void downloadImage(File output, URL source) {
 		
 		try {
+			/*
+			
+			System.out.println("path: " + output);
+			System.out.println("parent canWrite: " + output.getParentFile().canWrite());
+			System.out.println("isFile: " + output.isFile());
+			System.out.println("create new file: " + output.createNewFile());
+			*/
+			if(output.isFile()) return;
+			
 			// Get a stream to the image
 			InputStream in = source.openStream();
 			OutputStream out = new BufferedOutputStream(new FileOutputStream(output));
